Switch to repo-style checkout paths
Make kokoro_build.cmd and kokoro_build.sh work on a local machine where
the Kokoro variables haven't been set. Use "dev" as the build ID in
this case.

Switch to Visual Studio 2017, which doesn't set an environment variable
like VS140COMNTOOLS, so search for it at the default install path.

Switch to a prebuilt Python so it's easier to build it on a Windows
machine. e.g. My laptop has a "C:\Program Files\Python39" directory but
no "C:\python37".

Change-Id: I368d4f281820e7bc6c77ce5f64099d68000fe278