Migrate from centos 7 to rocky 8.
The centos 7 is finally EOL and the mirrors are gone, so this
container can't be built any more. Rocky claims to be the spiritual
successor to centos, so I've gone to that to try to minimize the
changes in the image. I still do have to go from 7 to 8 (the centos
and rocky version numbers actually are pretty comparable, since
they're both following the RHEL versioning) since rocky didn't exist
until 8.

I can't upgrade to anything newer than that because rocky 9 has a
glibc newer than Studio's minimum (2.34 vs Studio's 2.31 requirement),
whereas rocky 8 is still on 2.28.

Bug: None
Test: ./kokoro/kokoro_build.sh
Change-Id: I01589b6d74534182139ee5bd4b1bf24cfaf3f144